+++ src/main/java/org/reduct/client/ServerClient.java
@@ -2,11 +2,13 @@ package org.reduct.client;
 
 import org.reduct.common.exception.ReductException;
 import org.reduct.common.exception.ReductSDKException;
+import org.reduct.model.bucket.Buckets;
 import org.reduct.model.server.ServerInfo;
 
 public interface ServerClient {
 
    /**
+    * Get statistical information about the storage. You can use this method to get stats of the storage and check its version.
     * Attempts to retrieve the server information. Such as, version, bucket count, usage, uptime, oldest record,
     * latest record, and default settings.
     *
@@ -17,4 +19,6 @@ public interface ServerClient {
     * @throws ReductSDKException If, any client side error occurs.
     */
    ServerInfo getServerInfo() throws ReductException, ReductSDKException;
+
+   Buckets getList() throws ReductException, ReductSDKException;
 }