+++ src/main/java/org/reduct/utils/JsonUtils.java
@@ -1,0 +1,5 @@
+package org.reduct.utils;
+
+public class JsonUtils {
+
+}