+++ CHANGELOG.md
@@ -10,3 +10,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - BucketClient.createBucket() method to create a bucket [PR-10](https://github.com/reductstore/reduct-java/pull/10)
 - TokenClient.createToken() method to create a token [PR-4](https://github.com/reductstore/reduct-java/pull/4)
 - ServerClient.serverInfo() method to get server info [PR-3](https://github.com/reductstore/reduct-java/pull/3)
+- ServerClient.getList() method to get server info [PR-16](https://github.com/reductstore/reduct-java/pull/16)